Before I mess everything up with colliders and spheres
So, for now, the project is ok with building tunnels for everything, and subtracting the arch with all the void space from the solid wall. Now, next step would be to find an optimal way of taking one brick in the wall at a time, and get a list of all arch/vault bricks that are within a given radius, then do the boolean difference of this target brick with all the arch/vault bricks in the list, and do that for every brick. This should be solved by using colliders on arch/vault bricks, and some Physics.OverlapSphere code